chore(sample-data): t#3275 demo data
