+++ python/apps/taiga/src/taiga/base/utils/sample_data/media/logo.png
